refactor(runs): split open_asks.py along its responsibility seams (#673) (#685)
* refactor(runs): split open ask responsibilities (#673)

open_asks.py owned obligation creation, settlement, expiry and digest
projection in one 1104-line module. Split along the responsibility
seams: open_asks.py keeps creation and reopening, open_ask_settlement.py
takes the answer/route transitions, open_ask_digest.py takes expiry and
the straggler projection. All importers updated; no compatibility
re-exports, so each name has exactly one owner.

Closes #673


* refactor(runs): give the run-deferral lifecycle its own module (#673)

Folds the maintainability review of the first pass:

- run_deferrals.py owns deferral creation, its helpers, the expiry
  policy and the expiry transition, which previously spanned creation
  and digest. open_ask_digest.py is now read-only.
- Notice supersession moves to its canonical owner, obligation_notices.py,
  so no module imports a private name from another.
- assume_utc lands in brain/kernel/common/time.py, replacing the copy of
  the same rule in three modules. It is deliberately not ensure_utc:
  that one raises on naive datetimes where this one assumes UTC.


* refactor(cycles): give the timestamp rule one owner (#684) (#696)

* refactor(cycles): migrate _aware_utc callers onto the shared assume_utc helper (#684)

Removes the five private _aware_utc copies in brain/systems/cycles/ and points
their call sites at brain/kernel/common/time.assume_utc, added on the base
branch by #685.

Behaviour is unchanged: the fast suite passes at 4752 passed / 11 skipped with
no test expectation edited.

* refactor(cycles): give the timestamp rule one owner via assume_utc_optional (#684)

The first pass deleted the five private _aware_utc copies but inlined the same
rule at 24 call sites in two idioms, which moved the duplication rather than
removing it. Add one None-preserving helper beside assume_utc and route every
site through it.

assume_utc_optional exists because assume_utc(None) returns *now*, while these
callers hold nullable columns that need None to survive. Its docstring states
the difference from both assume_utc and ensure_utc, and a test pins all three
side by side.

The five originals were not identical: contracts.py and degradation.py converted
an aware non-UTC value to UTC, the other three returned it unchanged. Collapsed
to the converting rule after confirming the divergence is unreachable --
compute_next_run_at and _parse_one_time_run_at both end in .astimezone(utc), and
the only other sources are database reads.

---------
