+++ brain/systems/cycles/contracts.py
@@ -2,10 +2,11 @@
 
 from __future__ import annotations
 
-from datetime import datetime, timedelta, timezone
+from datetime import datetime, timedelta
 from typing import Any
 from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
 
+from brain.kernel.common.time import assume_utc_optional
 from brain.systems.cycles.common import (
     OFF_SLOT_MATERIAL_ALERT_RUN_KIND,
     SCHEDULED_CYCLE_ORIGIN,
@@ -46,19 +47,13 @@ CYCLE_RESULT_CONTRACT_REQUIRED_OUTPUTS_BY_RUN_KIND = {
 VALID_CYCLE_RUN_KINDS = frozenset(CYCLE_RESULT_CONTRACT_REQUIRED_OUTPUTS_BY_RUN_KIND)
 
 
-def _aware_utc(value: datetime | None) -> datetime | None:
-    if value is not None and value.tzinfo is None:
-        return value.replace(tzinfo=timezone.utc)
-    return value.astimezone(timezone.utc) if value is not None else None
-
-
 def _iso(value: datetime | None) -> str | None:
     return value.isoformat() if value is not None else None
 
 
 def cycle_scheduled_review_window(scheduled_for: datetime | None) -> dict[str, Any]:
     """Return the stable evidence window a scheduled cycle should inspect."""
-    end = _aware_utc(scheduled_for)
+    end = assume_utc_optional(scheduled_for)
     start = end - timedelta(hours=SCHEDULED_REVIEW_WINDOW_HOURS) if end else None
     return {
         "anchor": "cycle_run.scheduled_for",
@@ -163,8 +158,9 @@ def cycle_launch_receipt(
     except (TypeError, ValueError, ZoneInfoNotFoundError):
         timezone_name = "UTC"
         local_timezone = ZoneInfo("UTC")
-    if scheduled_for is not None:
-        local_scheduled_for = _aware_utc(scheduled_for).astimezone(
+    aware_scheduled_for = assume_utc_optional(scheduled_for)
+    if aware_scheduled_for is not None:
+        local_scheduled_for = aware_scheduled_for.astimezone(
             local_timezone
         ).isoformat()
     return {
@@ -173,7 +169,7 @@ def cycle_launch_receipt(
         "cycle_run_id": cycle_run_id,
         "origin": origin,
         "launch_context": launch,
-        "scheduled_for": _iso(_aware_utc(scheduled_for)),
+        "scheduled_for": _iso(aware_scheduled_for),
         "timezone": timezone_name,
         "local_scheduled_for": local_scheduled_for,
         "scheduled_review_window": cycle_scheduled_review_window(scheduled_for),
